escape paths if necessary
fix for #3
