+++ Cargo.toml
@@ -37,6 +37,7 @@ build-bin = [
 crossbeam-channel = "0.5.0"
 dunce = "1.0.1"
 globset = "0.4.6"
+grep-cli = "0.1.5"
 ignore = "0.4.16"
 log = "0.4.11"
 num_cpus = "1.13.0"
