+++ tests/common.rs
@@ -1,11 +1,25 @@
-mod test_dir;
-
-use test_dir::TestDir;
-
 /// quick-n-dirty any result type alias
-type AnyResult<T = (), E = Box<dyn std::error::Error>> = Result<T, E>;
+pub type AnyResult<T = (), E = Box<dyn std::error::Error>> = Result<T, E>;
 
-const MAX_LEN: usize = 256 * 1024;
+pub const MAX_LEN: usize = 256 * 1024;
+
+pub fn random_collection<T, I>(size: usize) -> I
+where
+    rand::distributions::Standard: rand::distributions::Distribution<T>,
+    I: std::iter::FromIterator<T>,
+{
+    use rand::Rng;
+    let mut rng = rand::thread_rng();
+    std::iter::repeat_with(|| rng.gen()).take(size).collect()
+}
+
+/// test shortcut
+pub fn find_dupes<P: AsRef<std::path::Path>>(path: &P) -> yadf::TreeBag<u64, yadf::path::Path> {
+    yadf::Yadf::builder()
+        .paths(&[path])
+        .build()
+        .scan::<seahash::SeaHasher>()
+}
 
 /// Test to be sure the sorting by hash only groups together files
 /// with the same contents.
@@ -24,283 +38,5 @@ fn sanity_check() {
             let contents = std::fs::read(&file).unwrap();
             assert_eq!(reference, contents, "comparing {:?} and {:?}", first, file);
         }
-    }
-}
-
-macro_rules! scope_name_iter {
-    () => {{
-        fn fxxfxxf() {}
-        fn type_name_of<T>(_: T) -> &'static str {
-            std::any::type_name::<T>()
-        }
-        type_name_of(fxxfxxf)
-            .split("::")
-            .take_while(|&segment| segment != "fxxfxxf")
-    }};
-}
-
-#[test]
-fn function_name() {
-    let fname = scope_name_iter!().collect::<Vec<_>>().join("::");
-    assert_eq!(fname, "common::function_name");
-}
-
-macro_rules! test_dir {
-    () => {{
-        ["target", "tests"]
-            .iter()
-            .copied()
-            .chain(scope_name_iter!())
-            .collect::<std::path::PathBuf>()
-    }};
-}
-
-#[test]
-fn dir_macro() {
-    let path = test_dir!();
-    #[cfg(windows)]
-    assert_eq!(path.to_str(), Some("target\\tests\\common\\dir_macro"));
-    #[cfg(not(windows))]
-    assert_eq!(path.to_str(), Some("target/tests/common/dir_macro"));
-}
-
-fn random_collection<T, I>(size: usize) -> I
-where
-    rand::distributions::Standard: rand::distributions::Distribution<T>,
-    I: std::iter::FromIterator<T>,
-{
-    use rand::Rng;
-    let mut rng = rand::thread_rng();
-    std::iter::repeat_with(|| rng.gen()).take(size).collect()
-}
-
-/// test shortcut
-fn find_dupes<P: AsRef<std::path::Path>>(path: &P) -> yadf::TreeBag<u64, std::path::PathBuf> {
-    yadf::Yadf::builder()
-        .paths(&[path])
-        .build()
-        .scan::<seahash::SeaHasher>()
-}
-
-#[test]
-fn identical_small_files() -> AnyResult {
-    let root = TestDir::new(test_dir!())?;
-    println!("{:?}", root.as_ref());
-    root.write_file("file1", b"aaa")?;
-    root.write_file("file2", b"aaa")?;
-    let counter = find_dupes(&root);
-    assert_eq!(counter.duplicates().iter().count(), 1);
-    assert_eq!(counter.as_tree().len(), 1);
-    Ok(())
-}
-
-#[test]
-fn identical_larger_files() -> AnyResult {
-    let root = TestDir::new(test_dir!())?;
-    let buffer: Vec<_> = random_collection(MAX_LEN * 3);
-    root.write_file("file1", &buffer)?;
-    root.write_file("file2", &buffer)?;
-    let counter = find_dupes(&root);
-    assert_eq!(counter.duplicates().iter().count(), 1);
-    assert_eq!(counter.as_tree().len(), 1);
-    Ok(())
-}
-
-#[test]
-fn files_differing_by_size() -> AnyResult {
-    let root = TestDir::new(test_dir!())?;
-    root.write_file("file1", b"aaaa")?;
-    root.write_file("file2", b"aaa")?;
-    let counter = find_dupes(&root);
-    assert_eq!(counter.duplicates().iter().count(), 0);
-    assert_eq!(counter.as_tree().len(), 2);
-    Ok(())
-}
-
-#[test]
-fn files_differing_by_prefix() -> AnyResult {
-    let root = TestDir::new(test_dir!())?;
-    root.write_file("file1", b"aaa")?;
-    root.write_file("file2", b"bbb")?;
-    let counter = find_dupes(&root);
-    assert_eq!(counter.duplicates().iter().count(), 0);
-    assert_eq!(counter.as_tree().len(), 2);
-    Ok(())
-}
-
-#[test]
-fn files_differing_by_suffix() -> AnyResult {
-    let root = TestDir::new(test_dir!())?;
-    let mut buffer1 = Vec::with_capacity(MAX_LEN * 3 + 4);
-    buffer1.extend_from_slice(&random_collection::<_, Vec<_>>(MAX_LEN * 3));
-    let mut buffer2 = buffer1.clone();
-    buffer1.extend_from_slice(b"suf1");
-    buffer2.extend_from_slice(b"suf2");
-    root.write_file("file1", &buffer1)?;
-    root.write_file("file2", &buffer2)?;
-    let counter = find_dupes(&root);
-    assert_eq!(counter.duplicates().iter().count(), 0);
-    assert_eq!(counter.as_tree().len(), 2);
-    Ok(())
-}
-
-#[test]
-fn files_differing_by_middle() -> AnyResult {
-    let root = TestDir::new(test_dir!())?;
-    let mut buffer1 = Vec::with_capacity(MAX_LEN * 2 + 4);
-    buffer1.extend_from_slice(&random_collection::<_, Vec<_>>(MAX_LEN));
-    let mut buffer2 = buffer1.clone();
-    buffer1.extend_from_slice(b"mid1");
-    buffer2.extend_from_slice(b"mid2");
-    let suffix = random_collection::<_, Vec<_>>(MAX_LEN);
-    buffer1.extend_from_slice(&suffix);
-    buffer2.extend_from_slice(&suffix);
-    root.write_file("file1", &buffer1)?;
-    root.write_file("file2", &buffer2)?;
-    let counter = find_dupes(&root);
-    assert_eq!(counter.duplicates().iter().count(), 0);
-    assert_eq!(counter.as_tree().len(), 2);
-    Ok(())
-}
-
-mod integration {
-    use super::test_dir::TestDir;
-    use super::{random_collection, AnyResult, MAX_LEN};
-    use predicates::{boolean::PredicateBooleanExt, str as predstr};
-
-    #[test]
-    fn trace_output() -> AnyResult {
-        let root = TestDir::new(test_dir!())?;
-        println!("{:?}", root.as_ref());
-        let bytes: Vec<_> = random_collection(MAX_LEN);
-        let file1 = root.write_file("file1", &bytes)?;
-        let file2 = root.write_file("file2", &bytes)?;
-        root.write_file("file3", &bytes[..4096])?;
-        root.write_file("file4", &bytes[..2048])?;
-        let expected = serde_json::to_string(&[[file1.to_string_lossy(), file2.to_string_lossy()]])
-            .unwrap()
-            + "\n";
-        assert_cmd::Command::cargo_bin(assert_cmd::crate_name!())?
-            .arg("-vvvv") // test stderr contains enough debug output
-            .args(&["--format", "json"])
-            .args(&["--algorithm", "seahash"])
-            .arg(root.as_ref())
-            .assert()
-            .success()
-            .stderr(
-                predstr::contains("Args {")
-                    .and(predstr::contains("Yadf {"))
-                    .and(predstr::contains("format: Json"))
-                    .and(predstr::contains("algorithm: SeaHash"))
-                    .and(predstr::contains("verbose: 4"))
-                    .and(predstr::contains(
-                        "found 3 possible duplicates after initial scan",
-                    ))
-                    .and(predstr::contains(
-                        "found 2 duplicates in 1 groups after checksumming",
-                    ))
-                    .and(predstr::contains("file1"))
-                    .and(predstr::contains("file2"))
-                    .and(predstr::contains("file3"))
-                    .and(predstr::contains("file4")),
-            )
-            .stdout(predstr::similar(expected).distance(2));
-        Ok(())
-    }
-
-    #[test]
-    fn regex() -> AnyResult {
-        let root = TestDir::new(test_dir!())?;
-        let bytes: Vec<_> = random_collection(4096);
-        let particular_1_name = root.write_file("particular_1_name", &bytes)?;
-        let particular_2_name = root.write_file("particular_2_name", &bytes)?;
-        root.write_file("not_particular_2_name", &bytes)?;
-        root.write_file("completely_different", &bytes)?;
-        let expected = [
-            particular_1_name.to_string_lossy(),
-            particular_2_name.to_string_lossy(),
-        ]
-        .join("\n")
-            + "\n";
-        assert_cmd::Command::cargo_bin(assert_cmd::crate_name!())?
-            .args(&["--regex", "^particular_\\d_name$"])
-            .arg(root.as_ref())
-            .assert()
-            .success()
-            .stderr(predstr::is_empty())
-            .stdout(predstr::similar(expected).distance(2));
-        Ok(())
-    }
-
-    #[test]
-    fn glob_pattern() -> AnyResult {
-        let root = TestDir::new(test_dir!())?;
-        let bytes: Vec<_> = random_collection(4096);
-        let particular_1_name = root.write_file("particular_1_name", &bytes)?;
-        let particular_2_name = root.write_file("particular_2_name", &bytes)?;
-        root.write_file("not_particular_2_name", &bytes)?;
-        root.write_file("completely_different", &bytes)?;
-        let expected = [
-            particular_1_name.to_string_lossy(),
-            particular_2_name.to_string_lossy(),
-        ]
-        .join("\n")
-            + "\n";
-        assert_cmd::Command::cargo_bin(assert_cmd::crate_name!())?
-            .args(&["--pattern", "particular*name"])
-            .arg(root.as_ref())
-            .assert()
-            .success()
-            .stderr(predstr::is_empty())
-            .stdout(predstr::similar(expected).distance(2));
-        Ok(())
-    }
-
-    #[test]
-    fn min_file_size() -> AnyResult {
-        let root = TestDir::new(test_dir!())?;
-        let bytes: Vec<_> = random_collection(4096);
-        let particular_1_name = root.write_file("particular_1_name", &bytes)?;
-        let particular_2_name = root.write_file("particular_2_name", &bytes)?;
-        root.write_file("not_particular_2_name", &bytes[..2048])?;
-        root.write_file("completely_different", &bytes[..2048])?;
-        let expected = [
-            particular_1_name.to_string_lossy(),
-            particular_2_name.to_string_lossy(),
-        ]
-        .join("\n")
-            + "\n";
-        assert_cmd::Command::cargo_bin(assert_cmd::crate_name!())?
-            .args(&["--min", "4K"])
-            .arg(root.as_ref())
-            .assert()
-            .success()
-            .stderr(predstr::is_empty())
-            .stdout(predstr::similar(expected).distance(2));
-        Ok(())
-    }
-
-    #[test]
-    fn max_file_size() -> AnyResult {
-        let root = TestDir::new(test_dir!())?;
-        let bytes: Vec<_> = random_collection(4096);
-        let particular_1_name = root.write_file("particular_1_name", &bytes[..1024])?;
-        let particular_2_name = root.write_file("particular_2_name", &bytes[..1024])?;
-        root.write_file("not_particular_2_name", &bytes)?;
-        root.write_file("completely_different", &bytes)?;
-        let expected = [
-            particular_1_name.to_string_lossy(),
-            particular_2_name.to_string_lossy(),
-        ]
-        .join("\n")
-            + "\n";
-        assert_cmd::Command::cargo_bin(assert_cmd::crate_name!())?
-            .args(&["--max", "2K"])
-            .arg(root.as_ref())
-            .assert()
-            .success()
-            .stderr(predstr::is_empty())
-            .stdout(predstr::similar(expected).distance(2));
-        Ok(())
     }
 }
